pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "invest-advisor-bot"
version = "0.2.0"
description = "Telegram investment advisor chatbot using market data, macro news, and technical trend analysis."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Codex" }]
dependencies = [
  "cachetools>=5.5.2",
  "httpx>=0.27.2",
  "loguru>=0.7.2",
  "lxml>=5.3.0",
  "numpy>=1.26.4",
  "pandas>=2.2.3",
  "psycopg[binary]>=3.2.9",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.6.1",
  "python-telegram-bot[job-queue]>=20.7",
  "yfinance>=0.2.54",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.6",
]
observability = [
  "mlflow>=2.17.2",
  "langfuse>=2.60.0",
]
analytics = [
  "duckdb>=1.2.0",
]
memory = [
  "qdrant-client>=1.11.0",
]
quality = [
  "great-expectations>=1.3.0",
]
feature-store = [
  "feast>=0.47.0",
]
evaluation = [
  "evidently>=0.7.0",
  "braintrust",
]
backtesting = [
  "vectorbt>=0.28.0",
]
adapters = [
  "openbb>=4.4.0",
]
execution = [
  "prefect>=3.0.0",
]
semantic-layer = [
  "dbt-core>=1.9.0",
]
microstructure = [
  "databento>=0.39.0",
]
streaming = [
  "clickhouse-connect>=0.8.15",
  "kafka-python-ng>=2.2.3",
  "redis>=6.0.0",
]
production = [
  "duckdb>=1.2.0",
  "qdrant-client>=1.11.0",
  "databento>=0.39.0",
  "clickhouse-connect>=0.8.15",
  "kafka-python-ng>=2.2.3",
  "redis>=6.0.0",
  "great-expectations>=1.3.0",
  "feast>=0.47.0",
  "evidently>=0.7.0",
  "braintrust",
  "mlflow>=2.17.2",
  "langfuse>=2.60.0",
  "vectorbt>=0.28.0",
  "prefect>=3.0.0",
  "dbt-core>=1.9.0",
]

[project.scripts]
invest-advisor-bot = "invest_advisor_bot.main:main"
invest-advisor-restore-backup = "invest_advisor_bot.backup_restore:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"
